refactor: enhanced react-app UI. + feat: setup ts path alias. + feat: added hooks folder. + fix: common Nextjs best practice.
