+++ packages/react-app/components/StorageContract.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
 import { Box, Button, Divider, Grid, Typography, Link } from "@mui/material";
 
-import { useInput } from ".";
+import { useInput } from "@/hooks/useInput";
 import { useContractKit } from "@celo-tools/use-contractkit";
 import { useEffect, useState } from "react";
-import { useSnackbar } from "notistack";
-import { truncateAddress } from "../utils";
+import { SnackbarAction, useSnackbar } from "notistack";
+import { truncateAddress } from "@/utils";
 import { Storage } from "../../hardhat/types/Storage";
 import { useQuery, gql } from "@apollo/client";
 
@@ -26,9 +26,9 @@ const QUERY = gql`
 
 export function StorageContract({ contractData }) {
   const { kit, address, network, performActions } = useContractKit();
-  const [storageValue, setStorageValue] = React.useState<string | null>(null);
+  const [storageValue, setStorageValue] = useState<string | null>(null);
   const [storageInput, setStorageInput] = useInput({ type: "text" });
-  const [contractLink, setContractLink] = React.useState<string | null>(null);
+  const [contractLink, setContractLink] = useState<string | null>(null);
   const { enqueueSnackbar, closeSnackbar } = useSnackbar();
   
   // Query the Graph endpoint specified in ../apollo-client.js 
@@ -59,14 +59,13 @@ export function StorageContract({ contractData }) {
 
         const result = await contract.methods
           .store(storageInput as string)
-          //@ts-ignore
           .send({ from: address, gasLimit });
 
         console.log(result);
 
         const variant = result.status == true ? "success" : "error";
         const url = `${network.explorer}/tx/${result.transactionHash}`;
-        const action = (key) => (
+        const action: SnackbarAction = (key) => (
           <>
             <Link href={url} target="_blank">
               View in Explorer
@@ -119,8 +118,8 @@ export function StorageContract({ contractData }) {
         <Typography variant="h6" component="div">
           Write Contract
         </Typography>
-        <Box sx={{ m: 1 }}>{setStorageInput}</Box>
-        <Button sx={{ m: 1 }} variant="contained" onClick={setStorage}>
+        <Box sx={{ m: 1, marginLeft: 0 }}>{setStorageInput}</Box>
+        <Button sx={{ m: 1, marginLeft: 0 }} variant="contained" onClick={setStorage}>
           Update Storage Contract
         </Button>
         <Divider component="div" sx={{ m: 1 }} />
@@ -128,10 +127,10 @@ export function StorageContract({ contractData }) {
         <Typography variant="h6" component="div">
           Read Contract
         </Typography>
-        <Typography sx={{ m: 1 }} component="div">
+        <Typography sx={{ m: 1, marginLeft: 0 }} component="div">
           Storage Contract Value: {storageValue}
         </Typography>
-        <Button sx={{ m: 1 }} variant="contained" onClick={getStorage}>
+        <Button sx={{ m: 1, marginLeft: 0 }} variant="contained" onClick={getStorage}>
           Read Storage Contract
         </Button>
       </Grid>
