+++ packages/react-app/components/index.ts
@@ -1,4 +1,3 @@
-export * from './useInput'
 export * from './ButtonAppBar'
 export * from './StorageContract'
 export * from './GreeterContract'
